+++ Job/Attribute.php
@@ -167,6 +167,8 @@ class Attribute extends Import
          * @var array $attribute
          */
         foreach ($attributes as $index => $attribute) {
+            $attribute['code'] = strtolower($attribute['code']);
+
             $this->entitiesHelper->insertDataFromApi($attribute, $this->getCode());
         }
         $index++;
@@ -325,6 +327,12 @@ class Attribute extends Import
         $connection = $this->entitiesHelper->getConnection();
         /** @var string $tmpTable */
         $tmpTable = $this->entitiesHelper->getTableName($this->getCode());
+
+        /** @var string $adminLang */
+        $adminLang = $this->storeHelper->getAdminLang();
+        /** @var string $adminLabelColumn */
+        $adminLabelColumn = sprintf('labels-%s', $adminLang);
+
         /** @var Select $import */
         $import = $connection->select()->from($tmpTable);
         /** @var \Zend_Db_Statement_Interface $query */
@@ -372,16 +380,10 @@ class Attribute extends Import
             // Prepare additional data for both tables.
 
             /* Retrieve default admin label */
-            /** @var array $stores */
-            $stores = $this->storeHelper->getStores();
             /** @var string $frontendLabel */
             $frontendLabel = __('Unknown');
-            if (isset($stores[0])) {
-                /** @var array $admin */
-                $admin = reset($stores[0]);
-                if (isset($row['labels-'.$admin['lang']])) {
-                    $frontendLabel = $row['labels-'.$admin['lang']];
-                }
+            if (!empty($row[$adminLabelColumn])) {
+                $frontendLabel = $row[$adminLabelColumn];
             }
 
             /* Retrieve attribute scope */