+++ Job/Product.php
@@ -70,11 +70,61 @@ class Product extends Import
      */
     protected $name = 'Product';
     /**
+     * Akeneo default association types, reformatted as column names
+     *
+     * @var string[] $associationTypes
+     */
+    protected $associationTypes = [
+        Link::LINK_TYPE_RELATED   => [
+            'SUBSTITUTION-products',
+            'SUBSTITUTION-product_models',
+        ],
+        Link::LINK_TYPE_UPSELL    => [
+            'UPSELL-products',
+            'UPSELL-product_models',
+        ],
+        Link::LINK_TYPE_CROSSSELL => [
+            'X_SELL-products',
+            'X_SELL-product_models',
+        ],
+    ];
+    /**
      * list of allowed type_id that can be imported
      *
      * @var string[]
      */
     protected $allowedTypeId = ['simple', 'virtual'];
+    /**
+     * List of column to exclude from attribute value setting
+     *
+     * @var string[]
+     */
+    protected $excludedColumns = [
+        '_entity_id',
+        '_is_new',
+        '_status',
+        '_type_id',
+        '_options_container',
+        '_tax_class_id',
+        '_attribute_set_id',
+        '_visibility',
+        '_children',
+        '_axis',
+        'identifier',
+        'sku',
+        'categories',
+        'family',
+        'groups',
+        'parent',
+        'enabled',
+        'created',
+        'updated',
+        'associations',
+        'PACK',
+        'SUBSTITUTION',
+        'UPSELL',
+        'X_SELL',
+    ];
     /**
      * This variable contains a ProductImportHelper
      *
@@ -532,7 +582,7 @@ class Product extends Import
         if ($connection->tableColumnExists($productTmpTable, 'parent')) {
             $groupColumn = 'parent';
         }
-        if ($connection->tableColumnExists($productTmpTable, 'groups') && !$groupColumn) {
+        if (!$groupColumn && $connection->tableColumnExists($productTmpTable, 'groups')) {
             $groupColumn = 'groups';
         }
         if ($groupColumn) {
@@ -547,7 +597,6 @@ class Product extends Import
                 ]
             );
         }
-
         $statusForConfigurables = $this->scopeConfig->getValue(ConfigHelper::ENABLE_CONFIGURABLE_IN_STORE) ? 1 : 2;
 
         /** @var array $data */
@@ -579,10 +628,26 @@ class Product extends Import
             $data['categories'] = 'e.categories';
         }
 
+        /** @var string[] $associationNames */
+        foreach ($this->associationTypes as $associationNames) {
+            if (empty($associationNames)) {
+                continue;
+            }
+            /** @var string $associationName */
+            foreach ($associationNames as $associationName) {
+                if (!empty($associationName) && $connection->tableColumnExists(
+                        $productModelTable,
+                        $associationName
+                    ) && $connection->tableColumnExists($productTmpTable, $associationName)) {
+                    $data[$associationName] = sprintf('v.%s', $associationName);
+                }
+            }
+        }
+
         /** @var string|array $additional */
         $additional = $this->scopeConfig->getValue(ConfigHelper::PRODUCT_CONFIGURABLE_ATTRIBUTES);
         $additional = $this->serializer->unserialize($additional);
-        if (!is_array($additional)) {
+        if (!\is_array($additional)) {
             $additional = [];
         }
 
@@ -633,15 +698,15 @@ class Product extends Import
             }
         }
 
-        /** @var Select $select */
-        $select = $connection->select()
+        /** @var Select $configurable */
+        $configurable = $connection->select()
             ->from(['e' => $productTmpTable], $data)
-            ->joinInner(['v' => $productModelTable],'e.' . $groupColumn . ' = v.code', [])
-            ->where('e.' . $groupColumn.' <> ""')
+            ->joinInner(['v' => $productModelTable], 'e.' . $groupColumn . ' = v.code', [])
+            ->where('e.' . $groupColumn . ' <> ""')
             ->group('e.' . $groupColumn);
 
         /** @var string $query */
-        $query = $connection->insertFromSelect($select, $configurableTmpTable, array_keys($data));
+        $query = $connection->insertFromSelect($configurable, $configurableTmpTable, array_keys($data));
         $connection->query($query);
 
         // Query tmp_pimgento_entities_low_level_configurables for entries without a parent,
@@ -930,24 +995,9 @@ class Product extends Import
         $columns = array_keys($connection->describeTable($tmpTable));
         /** @var string[] $except */
         $except = [
-            '_entity_id',
-            '_is_new',
-            '_status',
-            '_type_id',
-            '_options_container',
-            '_tax_class_id',
-            '_attribute_set_id',
-            '_visibility',
-            '_children',
-            '_axis',
-            'identifier',
-            'categories',
-            'family',
-            'groups',
-            'parent',
             'url_key',
-            'enabled',
         ];
+        $except = array_merge($except, $this->excludedColumns);
 
         /** @var string $column */
         foreach ($columns as $column) {
@@ -959,23 +1009,22 @@ class Product extends Import
                 continue;
             }
 
-            /** @var array|string $columnPrefix */
-            $columnPrefix = explode('-', $column);
-            $columnPrefix = reset($columnPrefix);
+            /** @var string[] $columnParts */
+            $columnParts = explode('-', $column, 2);
+            /** @var string $columnPrefix */
+            $columnPrefix = reset($columnParts);
+            $columnPrefix = sprintf('%s_', $columnPrefix);
             /** @var int $prefixLength */
-            $prefixLength = strlen($columnPrefix . '_') + 1;
+            $prefixLength = strlen($columnPrefix) + 1;
             /** @var string $entitiesTable */
             $entitiesTable = $this->entitiesHelper->getTable('pimgento_entities');
 
             // Sub select to increase performance versus FIND_IN_SET
             /** @var Select $subSelect */
-            $subSelect = $connection->select()
-                ->from(
-                    ['c' => $entitiesTable],
-                    ['code' => 'SUBSTRING(`c`.`code`, ' . $prefixLength . ')', 'entity_id' => 'c.entity_id']
-                )
-                ->where('c.code LIKE "' . $columnPrefix . '_%" ')
-                ->where('c.import = ?', 'option');
+            $subSelect = $connection->select()->from(
+                ['c' => $entitiesTable],
+                ['code' => sprintf('SUBSTRING(`c`.`code`, %s)', $prefixLength), 'entity_id' => 'c.entity_id']
+            )->where(sprintf('c.code LIKE "%s%s"', $columnPrefix, '%'))->where('c.import = ?', 'option');
 
             // if no option no need to continue process
             if (!$connection->query($subSelect)->rowCount()) {
@@ -1094,37 +1143,20 @@ class Product extends Import
         $connection = $this->entitiesHelper->getConnection();
         /** @var string $tmpTable */
         $tmpTable = $this->entitiesHelper->getTableName($this->getCode());
+        /** @var string[] $attributeScopeMapping */
+        $attributeScopeMapping = $this->entitiesHelper->getAttributeScopeMapping();
         /** @var array $stores */
         $stores = $this->storeHelper->getAllStores();
         /** @var string[] $columns */
         $columns = array_keys($connection->describeTable($tmpTable));
 
-        // Excluding irrelevant columns from processing.
-        /** @var string[] $except */
-        $except = [
-            '_entity_id',
-            '_is_new',
-            '_status',
-            '_type_id',
-            '_options_container',
-            '_tax_class_id',
-            '_attribute_set_id',
-            '_visibility',
-            '_children',
-            '_axis',
-            'sku',
-            'categories',
-            'family',
-            'groups',
-            'parent',
-            'enabled',
-        ];
+        $objectManager = \Magento\Framework\App\ObjectManager::getInstance();
+        $helper = $objectManager->get('\Magento\Directory\Helper\Data');
 
-        // Beginning to map data for later db write.
-        // The first item is a generic description.
-        // The following items are store-specific.
 
-        /** @var array $values */
+        /** @var string $adminBaseCurrency */
+        $adminBaseCurrency = $helper->getBaseCurrencyCode();
+        /** @var mixed[] $values */
         $values = [
             0 => [
                 'options_container' => '_options_container',
@@ -1137,7 +1169,7 @@ class Product extends Import
             $values[0]['status'] = '_status';
         }
 
-        /** @var array $taxClasses */
+        /** @var mixed[] $taxClasses */
         $taxClasses = $this->configHelper->getProductTaxClasses();
         if (count($taxClasses)) {
             foreach ($taxClasses as $storeId => $taxClassId) {
@@ -1147,39 +1179,65 @@ class Product extends Import
 
         /** @var string $column */
         foreach ($columns as $column) {
-            if (in_array($column, $except) || preg_match('/-unit/', $column)) {
+            /** @var string[] $columnParts */
+            $columnParts = explode('-', $column, 2);
+            /** @var string $columnPrefix */
+            $columnPrefix = $columnParts[0];
+
+            if (in_array($columnPrefix, $this->excludedColumns) || preg_match('/-unit/', $column)) {
                 continue;
             }
 
-            // Some columns may be relevant only to particular stores.
-            // Those have names patterned like so: prefix-suffix, where the suffix is the store name.
+            if (!isset($attributeScopeMapping[$columnPrefix])) {
+                // If no scope is found, attribute does not exist
+                continue;
+            }
 
-            /** @var array|string $columnPrefix */
-            $columnPrefix = explode('-', $column);
-            $columnPrefix = reset($columnPrefix);
+            if (empty($columnParts[1])) {
+                // No channel and no locale found: attribute scope naturally is Global
+                $values[0][$columnPrefix] = $column;
 
-            // For those columns, isolate the prefix and map it to the store in the values array.
+                continue;
+            }
 
-            /**
-             * @var string $suffix
-             * @var array $affected
-             */
-            foreach ($stores as $suffix => $affected) {
-                if (!preg_match('/^' . $columnPrefix . '-' . $suffix . '$/', $column)) {
+            /** @var int $scope */
+            $scope = (int)$attributeScopeMapping[$columnPrefix];
+            if ($scope === \Magento\Eav\Model\Entity\Attribute\ScopedAttributeInterface::SCOPE_GLOBAL && !empty($columnParts[1]) && $columnParts[1] === $adminBaseCurrency){
+                // This attribute has global scope with a suffix: it is a price with its currency
+                // If Price scope is set to Website, it will be processed afterwards as any website scoped attribute
+                $values[0][$columnPrefix] = $column;
+
+                continue;
+            }
+
+            /** @var string $columnSuffix */
+            $columnSuffix = $columnParts[1];
+            if (!isset($stores[$columnSuffix])) {
+                // No corresponding store found for this suffix
+                continue;
+            }
+
+            /** @var mixed[] $affectedStores */
+            $affectedStores = $stores[$columnSuffix];
+            /** @var mixed[] $store */
+            foreach ($affectedStores as $store) {
+                // Handle website scope
+                if ($scope === \Magento\Eav\Model\Entity\Attribute\ScopedAttributeInterface::SCOPE_WEBSITE && !$store['is_website_default']) {
                     continue;
                 }
 
-                /** @var array $store */
-                foreach ($affected as $store) {
-                    if (!isset($values[$store['store_id']])) {
-                        $values[$store['store_id']] = [];
-                    }
+                if ($scope === \Magento\Eav\Model\Entity\Attribute\ScopedAttributeInterface::SCOPE_STORE || empty($store['siblings'])) {
                     $values[$store['store_id']][$columnPrefix] = $column;
-                }
-            }
 
-            if (!isset($values[0][$columnPrefix])) {
-                $values[0][$columnPrefix] = $column;
+                    continue;
+                }
+
+                /** @var string[] $siblings */
+                $siblings = $store['siblings'];
+                /** @var string $storeId */
+                foreach ($siblings as $storeId) {
+                    $values[$storeId][$columnPrefix] = $column;
+                }
             }
         }
 
@@ -1188,7 +1246,7 @@ class Product extends Import
 
         /**
          * @var string $storeId
-         * @var array $data
+         * @var string[] $data
          */
         foreach ($values as $storeId => $data) {
             $this->entitiesHelper->setValues(
@@ -1414,12 +1472,12 @@ class Product extends Import
 
             /** @var Select $select */
             $select = $connection->select()->from(
-                    $tmpTable,
-                    [
-                        'product_id' => '_entity_id',
-                        'website_id' => new Expr($websiteId),
-                    ]
-                );
+                $tmpTable,
+                [
+                    'product_id' => '_entity_id',
+                    'website_id' => new Expr($websiteId),
+                ]
+            );
 
             $connection->query(
                 $connection->insertFromSelect(
@@ -1555,48 +1613,43 @@ class Product extends Import
         $linkTable = $this->entitiesHelper->getTable('catalog_product_link');
         /** @var string $linkAttributeTable */
         $linkAttributeTable = $this->entitiesHelper->getTable('catalog_product_link_attribute');
-        /** @var array $related */
+        /** @var mixed[] $related */
         $related = [];
 
         /** @var string $columnIdentifier */
         $columnIdentifier = $this->entitiesHelper->getColumnIdentifier($productsTable);
 
-        if ($connection->tableColumnExists($tmpTable, 'UPSELL-products')) {
-            $related[Link::LINK_TYPE_UPSELL][] = '`p`.`UPSELL-products`';
-        }
-        if ($connection->tableColumnExists($tmpTable, 'UPSELL-product_models')) {
-            $related[Link::LINK_TYPE_UPSELL][] = '`p`.`UPSELL-product_models`';
-        }
-
-        if ($connection->tableColumnExists($tmpTable, 'X_SELL-products')) {
-            $related[Link::LINK_TYPE_CROSSSELL][] = '`p`.`X_SELL-products`';
-        }
-        if ($connection->tableColumnExists($tmpTable, 'X_SELL-product_models')) {
-            $related[Link::LINK_TYPE_CROSSSELL][] = '`p`.`X_SELL-product_models`';
-        }
-
-        if ($connection->tableColumnExists($tmpTable, 'SUBSTITUTION-products')) {
-            $related[Link::LINK_TYPE_RELATED][] = '`p`.`SUBSTITUTION-products`';
-        }
-        if ($connection->tableColumnExists($tmpTable, 'SUBSTITUTION-product_models')) {
-            $related[Link::LINK_TYPE_RELATED][] = '`p`.`SUBSTITUTION-product_models`';
+        /** @var int $linkType */
+        /** @var string[] $associationNames */
+        foreach ($this->associationTypes as $linkType => $associationNames) {
+            if (empty($associationNames)) {
+                continue;
+            }
+            /** @var string $associationName */
+            foreach ($associationNames as $associationName) {
+                if (!empty($associationName) && $connection->tableColumnExists($tmpTable, $associationName)) {
+                    $related[$linkType][] = sprintf('`p`.`%s`', $associationName);
+                }
+            }
         }
 
+        /**
+         * @var int      $typeId
+         * @var string[] $columns
+         */
         foreach ($related as $typeId => $columns) {
-            $concat = 'CONCAT(' . join(',",",', $columns) . ')';
-            $select = $connection->select()
-                ->from(['c' => $entitiesTable], [])
-                ->joinInner(
-                    ['p' => $tmpTable],
-                    'FIND_IN_SET(`c`.`code`, ' . $concat . ') AND
-                        `c`.`import` = "' . $this->getCode() . '"',
-                    [
-                        'product_id'        => 'p._entity_id',
-                        'linked_product_id' => 'c.entity_id',
-                        'link_type_id'      => new Expr($typeId)
-                    ]
-                )
-                ->joinInner(['e' => $productsTable], 'c.entity_id = e.' . $columnIdentifier, []);
+            /** @var string $concat */
+            $concat = sprintf('CONCAT_WS(",", %s)', implode(', ', $columns));
+            /** @var \Magento\Framework\DB\Select $select */
+            $select = $connection->select()->from(['c' => $entitiesTable], [])->joinInner(
+                ['p' => $tmpTable],
+                sprintf('FIND_IN_SET(`c`.`code`, %s) AND `c`.`import` = "%s"', $concat, $this->getCode()),
+                [
+                    'product_id'        => 'p._entity_id',
+                    'linked_product_id' => 'c.entity_id',
+                    'link_type_id'      => new Expr($typeId),
+                ]
+            )->joinInner(['e' => $productsTable], sprintf('c.entity_id = e.%s', $columnIdentifier), []);
 
             /* Remove old link */
             $connection->delete(