+++ CHANGELOG.md
@@ -28,3 +28,39 @@ Fix documentation
 Reformat translations
 
 Fix localized product URL
+
+### 100.1.4 :
+Add PHP 7.2 compatibility
+
+Fix error on product_model table
+
+Add configuration to enable or disable product URL rewriting
+
+### 100.2.1 :
+Fix ACL for import jobs
+
+Fix family exclusion configuration that was impossible to set empty
+
+Fix configurable product association import
+
+Fix product website association from channel mapping
+
+Add configuration to set channel for admin values
+
+Add configuration to choose to update product URL rewrite or not
+
+Prevent price from being set to "null" if empty
+
+Add no-whitespace validation to Akeneo API credential fields
+
+### 100.2.2 :
+Fix option import if first attribute has no valid options
+
+Force attribute code to be lowercase
+
+Manage attribute scope on import
+
+Fix SQL error when no association column is found using advanced filter
+
+### 100.2.3 :
+Fix undefined offset on product import